debug fields not updating state on change
