+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/index.js":"1","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/App.js":"2","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/SideMenu.js":"3","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/Header.js":"4","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/PageHeader.js":"5","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/Employees.js":"6","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/EmployeeForm.js":"7"},{"size":220,"mtime":1610479793445,"results":"8","hashOfConfig":"9"},{"size":1025,"mtime":1610568473801,"results":"10","hashOfConfig":"9"},{"size":544,"mtime":1610484420732,"results":"11","hashOfConfig":"9"},{"size":2096,"mtime":1610551824192,"results":"12","hashOfConfig":"9"},{"size":1448,"mtime":1610492632521,"results":"13","hashOfConfig":"9"},{"size":793,"mtime":1610569285419,"results":"14","hashOfConfig":"9"},{"size":1524,"mtime":1610569489378,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"157wv0g",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"22","usedDeprecatedRules":"23"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/index.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/App.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/SideMenu.js",["32"],"import React from 'react'\nimport {makeStyles, withStyles} from '@material-ui/core'\n\n//withStyles & makeStyles\n\nconst style = {\n    sideMenu: {\n        display: 'flex',\n        flexDirection: 'column', \n        position: 'absolute', \n        left: '0px', \n        width: '320px', \n        height: '100%',\n        backgroundColor: '#253053'\n    }\n}\n\nconst SideMenu = (props) => {\n    const {classes} = props; \n    return (\n        <div className={classes.sideMenu}>\n            \n        </div>\n    )\n}\n\nexport default withStyles(style)(SideMenu)\n",["33","34"],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/Header.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/PageHeader.js",["35"],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/Employees.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/EmployeeForm.js",["36"],{"ruleId":"37","severity":1,"message":"38","line":2,"column":9,"nodeType":"39","messageId":"40","endLine":2,"endColumn":19},{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},{"ruleId":"37","severity":1,"message":"45","line":3,"column":10,"nodeType":"39","messageId":"40","endLine":3,"endColumn":24},{"ruleId":"37","severity":1,"message":"46","line":1,"column":26,"nodeType":"39","messageId":"40","endLine":1,"endColumn":35},"no-unused-vars","'makeStyles' is defined but never used.","Identifier","unusedVar","no-native-reassign",["47"],"no-negated-in-lhs",["48"],"'FullscreenExit' is defined but never used.","'useEffect' is defined but never used.","no-global-assign","no-unsafe-negation"]
+[{"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/index.js":"1","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/App.js":"2","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/SideMenu.js":"3","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/Header.js":"4","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/PageHeader.js":"5","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/Employees.js":"6","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/EmployeeForm.js":"7","/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/useForm.js":"8"},{"size":220,"mtime":1610479793445,"results":"9","hashOfConfig":"10"},{"size":1025,"mtime":1610568473801,"results":"11","hashOfConfig":"10"},{"size":532,"mtime":1610575081904,"results":"12","hashOfConfig":"10"},{"size":2096,"mtime":1610551824192,"results":"13","hashOfConfig":"10"},{"size":1395,"mtime":1610575056124,"results":"14","hashOfConfig":"10"},{"size":819,"mtime":1610574143916,"results":"15","hashOfConfig":"10"},{"size":1325,"mtime":1610577922744,"results":"16","hashOfConfig":"10"},{"size":765,"mtime":1610577629942,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"157wv0g",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":5,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/index.js",[],["35","36"],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/App.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/SideMenu.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/Header.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/PageHeader.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/Employees.js",[],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Pages/Employees/EmployeeForm.js",["37","38","39","40","41"],"/Users/jeff.woltjen.computer/react/material-ui-demo2/src/Components/useForm.js",[],{"ruleId":"42","replacedBy":"43"},{"ruleId":"44","replacedBy":"45"},{"ruleId":"46","severity":1,"message":"47","line":1,"column":16,"nodeType":"48","messageId":"49","endLine":1,"endColumn":24},{"ruleId":"46","severity":1,"message":"50","line":1,"column":26,"nodeType":"48","messageId":"49","endLine":1,"endColumn":35},{"ruleId":"46","severity":1,"message":"51","line":2,"column":26,"nodeType":"48","messageId":"49","endLine":2,"endColumn":36},{"ruleId":"46","severity":1,"message":"52","line":3,"column":18,"nodeType":"48","messageId":"49","endLine":3,"endColumn":22},{"ruleId":"46","severity":1,"message":"53","line":21,"column":9,"nodeType":"48","messageId":"49","endLine":21,"endColumn":18},"no-native-reassign",["54"],"no-negated-in-lhs",["55"],"no-unused-vars","'useState' is defined but never used.","Identifier","unusedVar","'useEffect' is defined but never used.","'makeStyles' is defined but never used.","'Form' is defined but never used.","'setValues' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
