+++ src/Components/SideMenu.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {makeStyles, withStyles} from '@material-ui/core'
+import {withStyles} from '@material-ui/core'
 
 //withStyles & makeStyles
