+++ src/Components/useForm.js
@@ -1,0 +1,40 @@
+import React, {useState} from 'react'; 
+import {makeStyles} from '@material-ui/core'
+
+export function useForm(initialValues) {
+
+    const [values, setValues] = useState(initialValues)
+
+      const handleInputChange = e => {
+        const {name, value} = e.target
+        setValues({
+            ...values,
+          [name]: value  
+        })
+    }
+    return {
+        values,
+        setValues,
+        handleInputChange
+    }
+}
+
+const useStyles = makeStyles(theme => ({
+    root: {
+        '& .MuiFormControl-root': {
+            width: '80%',
+            margin: theme.spacing(1)
+        }
+    }
+}))
+
+export function Form(props) {
+    const classes = useStyles(); 
+    return (
+        <Form className={classes.root}>
+{props.children}
+        </Form>
+    )
+}
+
+
